+++ @ornikar/storybook-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@1.2.2...@ornikar/storybook-config@1.2.3) (2020-04-30)
+
+**Note:** Version bump only for package @ornikar/storybook-config
+
+
+
+
+
 ## [1.2.2](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@1.2.1...@ornikar/storybook-config@1.2.2) (2020-04-21)
 
 **Note:** Version bump only for package @ornikar/storybook-config